+++ backend/backend.go
@@ -164,6 +164,12 @@ type authEventObject struct {
 	Result *AuthResultType `json:"result,omitempty"`
 }
 
+// NumberFormat can contain different decimal and group separators specified by the user to format numbers can be empty in case the user did not overwrite.
+type NumberFormat struct {
+	DecimalSeparator string `json:"decimal"`
+	GroupSeparator   string `json:"group"`
+}
+
 // Environment represents functionality where the implementation depends on the environment the app
 // runs in, e.g. Qt5/Mobile/webdev.
 type Environment interface {
@@ -186,6 +192,7 @@ type Environment interface {
 	// to return empty string or unsupported value, in which case the app will use
 	// English by default.
 	NativeLocale() string
+	NumberFormat() *NumberFormat
 	// Invoke a file picker dialog for the user to select a destination to store a file.
 	// `suggestedFilename` is the proposed/default destination.
 	// The function should return the empty string if the user aborted the process.