+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Bitcoin: show account details for the persisted receive address type by default
 - Add external block explorer links to used addresses
 - Settings search
+- Support macOS "Number Format" system setting
 
 ## v4.51.1
 - iOS: fix App Store submission by packaging Inter as TTF fonts